try out gql operation codegen
