+++ .gitignore
@@ -27,5 +27,6 @@ dist-ssr
 /build
 /src-tauri/src/graphql/generated.rs
 /src-tauri/src/graphql/datasource_impl.rs
-/src/graphql.d.ts
+/src/graphql/types.d.ts
+/src/graphql/operations.ts
 /graphql-schema.json
