+++ src/app/operations.graphql
@@ -1,0 +1,3 @@
+mutation initializeApp {
+  initializeApp
+}
